Rename books.py in models and resources to book.py, update README to document setting correct env variable for FLASK_ENV
